Fix polyfilltact.sh

+++ polyfilltact.sh
@@ -6,7 +6,18 @@ npm install yarn
 yarn install
 npm run gen || exit
 npm run build || exit
-rm -rf ../node_modules/@tact-lang/compiler/dist
-mv dist ../node_modules/@tact-lang/compiler/
-cd ..
-rm -rf tact-plus-plus
+# check if dir exists
+if [ -d "../node_modules/@tact-lang/compiler" ]; then
+  echo "NPM detected..."
+  rm -rf ../node_modules/@tact-lang/compiler/dist
+  mv dist ../node_modules/@tact-lang/compiler/
+  cd ..
+  rm -rf tact-plus-plus
+fi
+if [ -d "../node_modules/.pnpm/@tact-lang+compiler@1.5.1/node_modules/@tact-lang/compiler/dist" ]; then
+  echo "PNPM detected..."
+  rm -rf ../node_modules/.pnpm/@tact-lang+compiler@1.5.1/node_modules/@tact-lang/compiler/dist
+  mv dist ../node_modules/.pnpm/@tact-lang+compiler@1.5.1/node_modules/@tact-lang/compiler/
+  cd ..
+  rm -rf tact-plus-plus
+fi
